Added signout to results screen
