+++ client/app/.env
@@ -3,4 +3,4 @@ AMPLIFY_USER_POOL_ID=us-east-1_9Z4Cb3Pf0
 AMPLIFY_USER_POOL_WEB_CLIENT_ID=70mshnq7h8jdo5proj9eb2tont
 CONCERT_START_TIME=3000
 WEBSOCKET_API_URL=wss://v6u0asut4d.execute-api.us-east-1.amazonaws.com/Prod
-WEBSOCKET_API_STUB=false
+WEBSOCKET_API_STUB=STAGE_WAITING
