+++ client/app/src/screens/Welcome.js
@@ -3,17 +3,14 @@ import CountDown from 'react-native-countdown-component';
 import { Button } from 'react-native-elements';
 import EStyleSheet from 'react-native-extended-stylesheet';
 import { CONCERT_START_TIME } from 'react-native-dotenv';
-import { Auth } from 'aws-amplify';
 
 import WaitingScreen from './Waiting';
-import { store, actions } from '../state';
-import concertApi from '../api/concertApi';
 import COLORS from '../constants/Colors';
 import { STAGE_WAITING } from '../constants/Stages';
 
-const START_TIME = isNaN(parseInt(CONCERT_START_TIME, 10))
+const START_TIME = isNaN(Number(CONCERT_START_TIME))
   ? CONCERT_START_TIME
-  : new Date(Date.now() + parseInt(CONCERT_START_TIME, 10)).toISOString();
+  : new Date(Date.now() + Number(CONCERT_START_TIME)).toISOString();
 
 export default class WelcomeScreen extends Component {
   handleFinish = async () => {
@@ -22,21 +19,10 @@ export default class WelcomeScreen extends Component {
     this.forceUpdate();
   };
 
-  handleConnect = async () => {
-    const idToken = (await Auth.currentSession()).getIdToken();
-    concertApi.connect(idToken.getJwtToken());
-    store.dispatch(actions.setUID(idToken.payload['cognito:username']));
-  };
-
-  handleDisconnect = async () => {
-    concertApi.disconnect();
-    await Auth.signOut();
-    this.props.navigation.state.params.onStateChange('signIn');
-  };
-
   render() {
-    const { headerText, messageText, stageId } = this.props.navigation.state.params || {};
-    const isCountdown = !!CONCERT_START_TIME && !concertApi.isConnected();
+    const { headerText, messageText, stageId, isConnected, onConnect, onDisconnect } =
+      this.props.navigation.state.params || {};
+    const isCountdown = !!CONCERT_START_TIME && !isConnected;
     const until = (Date.parse(START_TIME) - Date.now()) / 1000;
     const isCountdownComplete = isCountdown && until <= 0;
 
@@ -46,14 +32,11 @@ export default class WelcomeScreen extends Component {
         messageText={!messageText && isCountdown ? '' : messageText}>
         {isCountdown &&
           (isCountdownComplete ? (
-            <Button
-              title="CONNECT"
-              onPress={this.handleConnect}
-              buttonStyle={styles.buttonConnect}
-            />
+            <Button title="CONNECT" onPress={onConnect} buttonStyle={styles.buttonConnect} />
           ) : (
             <CountDown
               digitStyle={styles.countdownDigit}
+              digitTxtStyle={styles.countdownDigitText}
               onFinish={this.handleFinish}
               size={24}
               timeLabelStyle={styles.countdownLabel}
@@ -61,11 +44,7 @@ export default class WelcomeScreen extends Component {
             />
           ))}
         {(isCountdown || stageId === STAGE_WAITING) && (
-          <Button
-            title="SIGNOUT"
-            onPress={this.handleDisconnect}
-            buttonStyle={styles.buttonDisconnect}
-          />
+          <Button title="SIGN OUT" onPress={onDisconnect} buttonStyle={styles.buttonDisconnect} />
         )}
       </WaitingScreen>
     );
@@ -82,6 +61,9 @@ const styles = EStyleSheet.create({
   },
   countdownDigit: {
     backgroundColor: COLORS.primaryOrange,
+  },
+  countdownDigitText: {
+    color: 'white',
   },
   countdownLabel: {
     color: 'white',
